+++ lib/flutter_iconpicker.dart
@@ -21,6 +21,10 @@ Future<IconPickerIcon?> showIconPicker(
   /// If non-null the icon picker highlights and scrolls to the selected icon
   IconPickerIcon? selectedIcon,
 
+  /// The background color for the [selectedIcon]
+  /// Defaults to `Theme.of(context).brightness == Brightness.dark ? Colors.grey[800] : Colors.grey[400]`
+  Color? selectedIconBackgroundColor,
+
   /// Defines if the searchbar will be
   /// shown above the icons
   bool showSearchBar = true,
@@ -163,6 +167,7 @@ Future<IconPickerIcon?> showIconPicker(
         context: context,
         builder: (BuildContext context) => FIPDefaultDialog(
           controller: controller,
+          selectedIconBackgroundColor: selectedIconBackgroundColor,
           showSearchBar: showSearchBar,
           adaptive: adaptiveDialog,
           showTooltips: showTooltips,
@@ -192,6 +197,7 @@ Future<IconPickerIcon?> showIconPicker(
           fullscreenDialog: true,
           builder: (context) => FIPDefaultDialog(
             controller: controller,
+            selectedIconBackgroundColor: selectedIconBackgroundColor,
             showSearchBar: showSearchBar,
             routedView: true,
             adaptive: adaptiveDialog,
@@ -223,6 +229,7 @@ Future<IconPickerIcon?> showIconPicker(
       context: context,
       builder: (BuildContext context) => FIPDefaultDialog(
         controller: controller,
+        selectedIconBackgroundColor: selectedIconBackgroundColor,
         showSearchBar: showSearchBar,
         showTooltips: showTooltips,
         barrierDismissible: barrierDismissible,